requirements-locust.txt
locust==2.31.8
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.1
